Now wingame waits until you press closebutton to leave the game.

+++ Nea Prototype/TopFrameManager.cs
@@ -54,7 +54,7 @@ namespace Nea_Prototype
         /// </summary>
         public void TryKillServer()
         {
-            if (ServerProcess != null)
+            if (ServerProcess != null && !ServerProcess.HasExited)
             {
                 serverProcess.Kill();
             }
